CKAN 2.9 support
