+++ ckanext/cloudstorage/templates/cloudstorage/snippets/cloudstorage-js_asset.html
@@ -1,0 +1,1 @@
+{% asset 'cloudstorage-js/main' %}
