+++ ckanext/cloudstorage/templates/cloudstorage/snippets/cloudstorage-js_resource.html
@@ -1,0 +1,1 @@
+{% resource 'cloudstorage-js/main' %}
