+++ ckanext/cloudstorage/templates/page.html
@@ -2,6 +2,7 @@
 
 {% block scripts %}
     {{ super() }}
-    {% resource 'cloudstorage-js/main' %}
+    {% set type = 'asset' if h.ckan_version() > '2.9' else 'resource' %}
+    {% include 'cloudstorage/snippets/cloudstorage-js_' ~ type ~ '.html' %}
 
 {% endblock %}
