+++ pyproject.toml
@@ -1,0 +1,3 @@
+[tool.black]
+line-length = 79
+include = '\.py$'
