+++ setup.py
@@ -25,7 +25,7 @@ setup(
         cloudstorage=ckanext.cloudstorage.plugin:CloudStoragePlugin
 
         [paste.paster_command]
-        cloudstorage=ckanext.cloudstorage.cli:PasterCommand
+        cloudstorage=ckanext.cloudstorage.commands:PasterCommand
         """
     ),
 )
